// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        InviteInfo: {
            /** Token Hash */
            token_hash: string;
            /** Short Code */
            short_code: string;
            /** Email */
            email: string;
            /** Role */
            role: string;
            /** Invited By */
            invited_by: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Expires At
             * Format: date-time
             */
            expires_at: string;
            /** Used At */
            used_at?: string | null;
            /** Used By */
            used_by?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}